* changes:
  feat(lib/smmuv3): derive StreamID and SMMU index for TDI identifier
  feat(lib/rmm_el3_ifc): refactor Root Complex list handling

+++ lib/smmuv3/src/fake_host/smmuv3_tlbi.c
@@ -14,7 +14,7 @@ int smmuv3_inv(void)
 	return 0;
 }
 
-int smmuv3_inv_entries(unsigned long smmu_idx, unsigned int vmid)
+int smmuv3_inv_entries(unsigned int smmu_idx, unsigned int vmid)
 {
 	(void)smmu_idx;
 	(void)vmid;
